Move ml react components to a folder
