+++ src/components/MachineLearning/ImageViews.tsx
