+++ src/components/MachineLearning/PeopleList.tsx
