+++ src/components/PhotoSwipe/PhotoSwipe.tsx
@@ -40,7 +40,10 @@ import { logError } from 'utils/sentry';
 
 import CloseIcon from 'components/icons/CloseIcon';
 import TickIcon from 'components/icons/TickIcon';
-import { PhotoPeopleList, UnidentifiedFaces } from 'components/PeopleList';
+import {
+    PhotoPeopleList,
+    UnidentifiedFaces,
+} from 'components/MachineLearning/PeopleList';
 import { FreeFlowText } from 'components/RecoveryKeyModal';
 import { Formik } from 'formik';
 import * as Yup from 'yup';
