+++ src/pages/ml-debug/index.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import dynamic from 'next/dynamic';
 
-const MLDebugWithNoSSR = dynamic(() => import('components/MlDebug'), {
-    ssr: false,
-});
+const MLDebugWithNoSSR = dynamic(
+    () => import('components/MachineLearning/MlDebug'),
+    {
+        ssr: false,
+    }
+);
 
 export default function MLDebug() {
     return (
